requirements_3090.txt
absl-py==0.15.0
appdirs==1.4.4
audioread==2.1.9
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cycler==0.11.0
Cython==0.29.24
decorator==4.4.2
dlib==19.22.1
einops==0.3.2
future==0.18.2
g2p-en==2.1.0
google==3.0.0
grpcio==1.42.0
h5py==2.8.0
horology==1.2.0
idna==3.3
imageio==2.10.1
imageio-ffmpeg==0.4.5
importlib-metadata==4.8.1
joblib==1.1.0
kiwisolver==1.3.2
librosa==0.8.0
llvmlite==0.31.0
Markdown==3.3.4
matplotlib==3.4.3
miditoolkit==0.1.7
moviepy==1.0.3
numba==0.48.0
numpy==1.20.0
opencv-python==4.5.4.58
packaging==21.2
pandas==1.3.4
Pillow==8.4.0
pooch==1.5.2
praat-parselmouth==0.3.3
proglog==0.1.9
protobuf==3.19.1
pycparser==2.20
pycwt==0.3.0a22
pydub==0.25.1
pyloudnorm==0.1.0
pyparsing==2.4.7
pypinyin==0.43.0
python-dateutil==2.8.2
pytorch-lightning==0.7.1
pytorch-ssim==0.1
pytz==2021.3
pyworld==0.3.0
PyYAML==6.0
requests==2.26.0
resampy==0.2.2
Resemblyzer==0.1.1.dev0
scikit-image==0.16.2
scikit-learn==0.22
scipy==1.3.0
six==1.16.0
sklearn==0.0
SoundFile==0.10.3.post1
soupsieve==2.3
sympy==1.9
tensorboard==1.15.0
tensorboardX==2.4
test-tube==0.7.5
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
uuid==1.30
webrtcvad==2.0.10
Werkzeug==2.0.2
zipp==3.6.0
